+++ ktlint-rule-engine/src/main/kotlin/com/pinterest/ktlint/rule/engine/internal/EditorConfigFinder.kt
@@ -19,7 +19,9 @@ import kotlin.system.measureTimeMillis
 
 private val LOGGER = KotlinLogging.logger {}.initKtLintKLogger()
 
-internal class EditorConfigFinder(private val editorConfigLoaderEc4j: EditorConfigLoaderEc4j) {
+internal class EditorConfigFinder(
+    private val editorConfigLoaderEc4j: EditorConfigLoaderEc4j,
+) {
     /**
      * Finds all relevant ".editorconfig" files for the given path.
      */