+++ ktlint-rule-engine/src/test/kotlin/com/pinterest/ktlint/rule/engine/internal/rules/KtlintSuppressionRuleTest.kt
@@ -1366,4 +1366,6 @@ class KtlintSuppressionRuleTest {
     }
 }
 
-private class DummyRule(id: String) : Rule(RuleId(id), About())
+private class DummyRule(
+    id: String,
+) : Rule(RuleId(id), About())